+++ KarhooSDK/Network/Common/DecodableData.swift
@@ -14,8 +14,4 @@ public struct DecodableData: KarhooCodableModel {
     public init(data: Data) {
         self.data = data
     }
-
-    public func encode(to encoder: Encoder) throws {
-
-    }
 }
