+++ KarhooSDK/Service/Payment/AdyenPayments/AdyenPaymentsInteractor.swift
@@ -9,6 +9,5 @@
 import Foundation
 
 protocol AdyenPaymentsInteractor: KarhooExecutable {
-
     func set(request: AdyenPaymentsRequest)
 }
